requirements.txt
# requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
requests
assemblyai
google-generativeai
websockets